refactor: tablegroup 이동
