+++ src/main/java/kitchenpos/exception/table/OrderTableIsInOtherTableGroupBadRequest.java
@@ -2,10 +2,10 @@ package kitchenpos.exception.table;
 
 import kitchenpos.exception.common.BadRequestException;
 
-public class OrderTableIsInTableGroupBadRequest extends BadRequestException {
+public class OrderTableIsInOtherTableGroupBadRequest extends BadRequestException {
     private static final String resourceName = "주문 테이블";
 
-    public OrderTableIsInTableGroupBadRequest(final long id) {
+    public OrderTableIsInOtherTableGroupBadRequest(final long id) {
         super(resourceName, id);
     }
 }
