+++ src/test/java/kitchenpos/application/integration/ApplicationIntegrationTest.java
@@ -4,8 +4,8 @@ import kitchenpos.application.mengroup.MenuGroupService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.order.OrderService;
 import kitchenpos.application.product.ProductService;
-import kitchenpos.application.table.TableGroupService;
 import kitchenpos.application.table.TableService;
+import kitchenpos.application.tablegroup.TableGroupService;
 import kitchenpos.domain.common.Money;
 import kitchenpos.dto.menu.CreateMenuRequest;
 import kitchenpos.dto.menu.MenuProductRequest;
